chroma-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
chromadb>=0.5.0
google-generativeai>=0.8.0
python-multipart>=0.0.12
pillow>=11.0.0
pypdf>=4.0.0
python-dotenv>=1.0.0
